refactor(lowering): extract program declaration helpers
